Many Changes
created a 'mod_assets' folder for items that will NOT be put in manually
using the editor and should be loaded automatically. Moved the new
'emtpy.wav' file to the proper folder.
Started a GUI script for handling shop and other instances.
Also, added a proof of concept for party interaction with NPCs: the
'monsters.lua' file in the 'mod_assets' folder holds the code.