Royal Archives Concept added
This is what first came to my mind when talking about the Royal
Archives. Note that I absolutely love what we already have, but I
thought it was worth posting if we need further ideas.

That is from Dark Souls if you were wondering. I was unable to find a
better screenshot online.